Release: v2.61.0

+++ src/detection/gpu/gpu_obsd.c
@@ -8,7 +8,7 @@
 #include <dev/pci/pcidevs.h>
 #include <sys/pciio.h>
 
-static inline int pciReadConf(int fd, uint32_t bus, uint32_t device, uint32_t func, uint32_t reg, uint32_t* result)
+static inline int pciReadConf(int fd, uint8_t bus, uint8_t device, uint8_t func, uint8_t reg, uint32_t* result)
 {
     struct pci_io bdfr = {
         .pi_sel = {
@@ -34,21 +34,21 @@ const char* ffDetectGPUImpl(FF_MAYBE_UNUSED const FFGPUOptions* options, FFlist*
     if (pcifd < 0)
         return "open(\"/dev/pci0\", O_RDONLY | O_CLOEXEC) failed";
 
-    for (uint32_t bus = 0; bus <= 255; bus++)
+    for (uint16_t bus = 0; bus <= 255; bus++)
     {
-        for (uint32_t dev = 0; dev < 32; dev++)
+        for (uint8_t dev = 0; dev < 32; dev++)
         {
-            uint32_t maxfuncs = 0;
-            for (uint32_t func = 0; func <= maxfuncs; func++)
+            uint8_t maxfuncs = 0;
+            for (uint8_t func = 0; func <= maxfuncs; func++)
             {
                 uint32_t pciid, pciclass;
-                if (pciReadConf(pcifd, bus, dev, func, PCI_ID_REG, &pciid) != 0)
+                if (pciReadConf(pcifd, (uint8_t) bus, dev, func, PCI_ID_REG, &pciid) != 0)
                     continue;
 
                 if (PCI_VENDOR(pciid) == PCI_VENDOR_INVALID || PCI_VENDOR(pciid) == 0)
                     continue;
 
-                if (pciReadConf(pcifd, bus, dev, func, PCI_CLASS_REG, &pciclass) != 0)
+                if (pciReadConf(pcifd, (uint8_t) bus, dev, func, PCI_CLASS_REG, &pciclass) != 0)
                     continue;
 
                 if (func == 0)
@@ -56,7 +56,7 @@ const char* ffDetectGPUImpl(FF_MAYBE_UNUSED const FFGPUOptions* options, FFlist*
                     // For some reason, pciReadConf returns success even for non-existing devices.
                     // So we need to check for `PCI_VENDOR(pciid) == PCI_VENDOR_INVALID` above to filter them out.
                     uint32_t bhlcr;
-                    if (pciReadConf(pcifd, bus, dev, 0, PCI_BHLC_REG, &bhlcr) != 0)
+                    if (pciReadConf(pcifd, (uint8_t) bus, dev, 0, PCI_BHLC_REG, &bhlcr) != 0)
                         continue;
 
                     if (PCI_HDRTYPE_MULTIFN(bhlcr)) maxfuncs = 7;
